Split message screen in two classes.
